# VoteApiCore21/PielievinGerasApiVoter/PielievinGerasApiVoter/Models/Candidate.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace PielievinGerasApiVoter.Models
{
    public class Candidate
    {
        public int Id { get; set; }
        public string Name { get; set; }
    }
}