# VoteApiCore21/temp/2.12.2018/PielievinGerasApiVoter/PielievinGerasApiVoter/Models/Candidate.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace PielievinGerasApiVoter.Models
{
    public class Candidate
    {
        

        public int Id { get; set; }
        public string Name { get; set; }
    }
}