# VoteApiCore21/temp/30.11.2018/VoterClient/VoterClientPielievinGeras/VoterClientPielievinGeras/Models/Candidate.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;

namespace VoterClientPielievinGeras.Models
{
    public class Candidate
    {
        public int Id { get; set; }
        public string Name { get; set; }
    }
}