# VoteApiCore21/temp/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Voter Service API"},"paths":{"/api/Candidates":{"get":{"tags":["Candidates"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Candidate"}}}}},"post":{"tags":["Candidates"],"operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"value","in":"body","required":false,"schema":{"$ref":"#/definitions/Candidate"}}],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Candidate"}}}}}},"/api/Candidates/{newName}":{"put":{"tags":["Candidates"],"operationId":"Put","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"value","in":"body","required":false,"schema":{"$ref":"#/definitions/Candidate"}},{"name":"newName","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/api/Candidates/GetDeleteLast":{"delete":{"tags":["Candidates"],"operationId":"GetDeleteLast","consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}}},"/api/Candidates/GetDeleteAll":{"delete":{"tags":["Candidates"],"operationId":"GetDeleteAll","consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}}},"/api/Votes":{"get":{"tags":["Votes"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Result"}}}}},"post":{"tags":["Votes"],"operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"value","in":"body","required":false,"schema":{"$ref":"#/definitions/Vote"}}],"responses":{"201":{"description":"Success","schema":{"type":"string"}},"400":{"description":"Bad Request"}}}}},"definitions":{"Candidate":{"type":"object","properties":{"id":{"format":"int32","type":"integer"},"name":{"type":"string"}}},"Result":{"type":"object","properties":{"candidate_id":{"format":"int32","type":"integer"},"count":{"format":"int32","type":"integer"}}},"Vote":{"type":"object","properties":{"id":{"format":"int32","type":"integer"},"candidate_id":{"format":"int32","type":"integer"},"user_id":{"format":"int32","type":"integer"}}}}}